manual rebase LPC stuff to 'main'

+++ projects/epc/acceptance-tests/src/mock/DspHostDualTester.cpp
@@ -232,10 +232,10 @@ DspHostDualTester::StereoOutput DspHostDualTester::scanOutput(const std::vector<
   return result;
 }
 
-double DspHostDualTester::encodeUnisonVoice(const unsigned int _unison, const unsigned int _polyphony)
+float DspHostDualTester::encodeUnisonVoice(const unsigned int _unison, const unsigned int _polyphony)
 {
   nltools_assertAlways(_unison > 0 && _unison <= _polyphony);
-  return static_cast<double>(_unison - 1) / static_cast<double>(_polyphony - 1);
+  return static_cast<float>(_unison - 1) / static_cast<float>(_polyphony - 1);
 }
 
 void DspHostDualTester::setSplit(VoiceGroup vg, float pos)
@@ -257,12 +257,12 @@ void DspHostDualTester::applyMalformedSplitPreset(const DspHostDualTester::Malfo
                                                   float split)
 {
   nltools::msg::SplitPresetMessage msg;
-  msg.mono[0].monoEnable.controlPosition = static_cast<double>(_partI.m_mono);
-  msg.mono[1].monoEnable.controlPosition = static_cast<double>(_partII.m_mono);
+  msg.mono[0].monoEnable.controlPosition = static_cast<float>(_partI.m_mono);
+  msg.mono[1].monoEnable.controlPosition = static_cast<float>(_partII.m_mono);
   msg.unison[0].unisonVoices.controlPosition = encodeUnisonVoice(_partI.m_unison, C15::Config::local_polyphony);
   msg.unison[1].unisonVoices.controlPosition = encodeUnisonVoice(_partII.m_unison, C15::Config::local_polyphony);
-  msg.splitpoint[0].controlPosition = static_cast<double>(split);
-  msg.splitpoint[1].controlPosition = static_cast<double>(split) + (1.0 / 61.0);
+  msg.splitpoint[0].controlPosition = split;
+  msg.splitpoint[1].controlPosition = split + (1.0 / 61.0);
   // propagate message
   m_host->onPresetMessage(msg);
 }
